Moved constants to appropriate dir
